Move VerletTraversalInterface to own file
